+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/Moim.java
@@ -19,7 +19,7 @@ import com.namo.spring.application.external.domain.group.domain.constant.MoimSta
 
 import com.namo.spring.application.external.domain.user.domain.User;
 
-import com.namo.spring.application.external.global.common.entity.BaseTimeEntity;
+import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
