+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/MoimMemoLocationAndUser.java
@@ -12,7 +12,7 @@ import jakarta.persistence.Table;
 
 import com.namo.spring.application.external.domain.user.domain.User;
 
-import com.namo.spring.application.external.global.common.entity.BaseTimeEntity;
+import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
