+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/MoimScheduleAndUser.java
@@ -22,7 +22,7 @@ import com.namo.spring.application.external.domain.individual.domain.Category;
 
 import com.namo.spring.application.external.domain.user.domain.User;
 
-import com.namo.spring.application.external.global.common.entity.BaseTimeEntity;
+import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
