+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/domain/Category.java
@@ -14,7 +14,7 @@ import jakarta.persistence.ManyToOne;
 import com.namo.spring.application.external.domain.individual.domain.constant.CategoryKind;
 import com.namo.spring.application.external.domain.individual.domain.constant.CategoryStatus;
 import com.namo.spring.application.external.domain.user.domain.User;
-import com.namo.spring.application.external.global.common.entity.BaseTimeEntity;
+import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
