+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/domain/Schedule.java
@@ -21,7 +21,7 @@ import com.namo.spring.application.external.domain.individual.domain.constant.Pe
 import com.namo.spring.core.common.exception.IndividualException;
 import com.namo.spring.application.external.domain.user.domain.User;
 
-import com.namo.spring.application.external.global.common.entity.BaseTimeEntity;
+import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
 import com.namo.spring.core.common.code.status.ErrorStatus;
 
 import lombok.AccessLevel;
