+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/common/model/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.global.common.entity;
+package com.namo.spring.db.mysql.common.model;
 
 import java.time.LocalDateTime;
